tpp/worker/requirements.txt
httpx==0.27.0
beautifulsoup4==4.12.3
lxml==5.3.0
rapidfuzz==3.9.6
python-slugify==8.0.4
